Add `ScheduledThreadPool::with_name_and_drop_behavior` method that creates a `ScheduledThreadPool` with the provided name, number of threads, and the desired behavior in relation to pending scheduled executions when the pool is dropped.
This on drop behavior is specified using the new `OnPoolDropBehavior` enum.

The existing behavior (and the default) is specified as
`CompletePendingScheduled`. This means that all pending scheduled executions
will be run, but periodic actions will not be rescheduled once these have
completed.

The other two options are `DiscardPendingScheduled` and
`RunPendingScheduledImmediately`. As the names suggest,
`DiscardPendingScheduled` will discard any pending scheduled executions and
`RunPendingScheduledImmediately` will immediately run any pending scheduled
executions.

The `RunPendingScheduledImmediately` option is useful in scenarios where you
know that any pending scheduled executions must be run but you don't want to
have to wait for them to be run at the time they were originally scheduled to be
run at. You need to be careful with this though because if you have multiple
pending scheduled executions then there's no guarantee as to what order they
will be run in. They will be popped from the queue in the expected order but
because they all get run immediately (or as immediately as is possible, given
the number of workers) then there's no guarantee as to order of execution. This
isn't, however, any different to two pending scheduled executions that were
scheduled to be run at the same time. They would similarly be popped from the
queue in a deterministic order but the threads that run the jobs are then in
control of exactly when the job is run.